fix: windows compilation errors
